Remove helpers.ts, sort into different files
